% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDiff.R
\name{checkDiff}
\alias{checkDiff}
\title{checkDiff Dense Tensor only. DO NOT USE FOR HUGE TENSOR.}
\usage{
checkDiff(Xnew, Xold)
}
\arguments{
\item{Xnew}{WIP}

\item{Xold}{WIP}
}
\value{
WIP
}
\description{
checkDiff Dense Tensor only. DO NOT USE FOR HUGE TENSOR.
}

