% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSTDFixedNum.R
\name{FSTDFixedFNum}
\alias{FSTDFixedFNum}
\title{FSTD}
\usage{
FSTDFixedFNum(Y, k)
}
\arguments{
\item{Y}{saimple_sparse_array (3rd-order tensor only)}

\item{k}{numeric. number of fibers to be picked up.}
}
\value{
list. $g is core tensor array. $As is feature matrices' list.
}
\description{
FSTD
}

