% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kModeProduct.R
\name{kModeProduct}
\alias{kModeProduct}
\title{kModeProduct}
\usage{
kModeProduct(tnsr, mat, m)
}
\arguments{
\item{tnsr}{simple_sparse_array}

\item{mat}{simple_sparse_array or simple_triplet_matrix or matrix. nrow(mat) should be equal to dim(tnsr)[m]}

\item{m}{mode}
}
\value{
product. simple_sparse_array.
}
\description{
kModeProduct
}

