% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refold.R
\name{Refold}
\alias{Refold}
\title{Refold}
\usage{
Refold(mat, dimToRecover, mode)
}
\arguments{
\item{mat}{kModeUnfolded 2-dimensional simple_sparse_array (not simple_triplet_matrix)}

\item{dimToRecover}{size of tensor to make}

\item{mode}{mode of unfold}
}
\value{
folded tensor (simple_sparse_array)
}
\description{
Refold
}

