% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartsvd.R
\name{smartsvd}
\alias{smartsvd}
\title{SVD with irlba}
\usage{
smartsvd(mat, rank = 10, nu = 3, nv = 3)
}
\arguments{
\item{mat}{WIP}

\item{rank}{WIP}

\item{nu}{WIP}

\item{nv}{WIP}
}
\value{
WIP
}
\description{
SVD with irlba
}

