\name{TensorSlam-package}
\alias{TensorSlam-package}
\alias{TensorSlam}
\docType{package}
\title{
\packageTitle{TensorSlam}
}
\description{
\packageDescription{TensorSlam}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TensorSlam}
\packageIndices{TensorSlam}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{TensorSlam}

Maintainer: \packageMaintainer{TensorSlam}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
