% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOSVD.R
\name{HOSVD}
\alias{HOSVD}
\title{Calc HOSVD}
\usage{
HOSVD(tnsr, ranks = ceiling(dim(tnsr)/10), doMC = T)
}
\arguments{
\item{tnsr}{Tensor. simple_sparse_array}

\item{ranks}{ranks}

\item{doMC}{Use doMC for parallel calculation or not.}
}
\value{
list. $g is a core tensor (simple_sparse_array) $U is a list that contains matrix (matrix)
}
\description{
Calc HOSVD
}

