% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOSVD.R
\name{HOSVD}
\alias{HOSVD}
\title{Calc HOSVD}
\usage{
HOSVD(tnsr, ranks = ceiling(dim(tnsr)/10), doMC = T)
}
\arguments{
\item{tnsr}{WIP}

\item{ranks}{WIP}

\item{doMC}{WIP}
}
\value{
WIP
}
\description{
Calc HOSVD
}

