% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kModeUnfold.R
\name{kModeUnfold}
\alias{kModeUnfold}
\title{kModeUnfold}
\usage{
kModeUnfold(tnsr, m)
}
\arguments{
\item{tnsr}{simple_sparse_array}

\item{m}{mode index to unfold}
}
\value{
2-dimensional simple_sparse_array (not simple_triplet_matrix)
}
\description{
kModeUnfold
}

