% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUNTD.R
\name{MUNTD}
\alias{MUNTD}
\title{Calc Nonnegative-Tucker-Decomposition}
\usage{
MUNTD(y, core_dims = ceiling(dim(y)/5), lra_ranks = core_dims,
  method = "FSTD", convDebug = F)
}
\arguments{
\item{y}{Input tensor. simple_spase_array}

\item{core_dims}{Number of Features/Size of core tensor. vector}

\item{lra_ranks}{Ranks for Low Rank Approximation. numeric}

\item{method}{Methods for Low Rank Approximation. 'FSTD'/'HOSVD'}

\item{convDebug}{WIP}
}
\value{
list. $g:core tensor (array) $As:Feature matrices (list,each matrix is matrix-class)
}
\description{
Calc Nonnegative-Tucker-Decomposition
}

