% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{WIP}
\alias{FSTD}
\alias{WIP}
\alias{inew}
\title{Other Modules}
\usage{
inew(C, old_ind)

FSTD(Y, k, eps = NULL)
}
\arguments{
\item{C}{WIP}

\item{old_ind}{WIP}

\item{Y}{WIP}

\item{k}{WIP}

\item{eps}{WIP}
}
\description{
Other Modules
}

